requirements.txt
pandas
httpx
python-dotenv
fastapi
loguru
uvicorn
bs4
openai
pygame
opencv-python
python-multipart
Pillow
pyyaml
opencv-contrib-python
